+++ examples/empty_partial.js
@@ -1,0 +1,3 @@
+var partial_context = {
+  foo: 1
+};