+++ mustache.js
@@ -282,7 +282,7 @@ var Mustache = function() {
 
   return({
     name: "mustache.js",
-    version: "0.2.3-dev",
+    version: "0.2.3",
 
     /*
       Turns a template and view into HTML