+++ CHANGES.md
@@ -1,7 +1,9 @@
 # mustache.js Changes
 
 ## 0.2.3 (??-??-????)
+* Better error message for missing partials.
 * Added more robust type detection.
+
 
 ## 0.2.2 (11-02-2010)